Ticket #22399: - Added nuke_script_path_step to info.yml - Added plate_published_file_type to info.yml - Added nuke_script_toolkit_write_nodes to info.yml
Introduced setting to specify the toolkit write nodes to make available to Hiero.
The UI in Hiero gets built independent of the tasks, so all the template paths
to the nuke scripts have no way to be resolved.  Instead of having a dependency
on the shot task environment implied by the export, just be explicit about the
node profiles to use.

Introduce a new GUI element to allow the export to toggle the different nodes
on and off.