#33014 Handles lack of thumbnail source more gracefully.
In the event that there's no rendered data on disk for the track item being exported, we won't have any thumbnail material available. In that situation, it's best to continue on with the export rather than let an unhelpful RuntimeError bubble up. This way, a more sane error message from Hiero/Nuke Studio will make its way to the user concerning the lack of frames on disk should that be a problem for the exporters.