For #35432 handle intermittent Windows issue removing temp thumbnail
Windows occasionally holds on to the temporary thumbnail file longer than it needs to which can cause an error when we try and remove it after uploading it to Shotgun: "[Error 32] The process cannot access the file because it is being used by another process". This fix catches the error and waits 1 second for the process to let go of the file before trying again. If for some reason the file still cannot be removed, we raise the exception.