#35773 Adds support for new Cuts schema to hiero export
When the export is initiated via the Process as Shotgun Shots processor, a
pre-process step now introspects the tasks to be performed and creates the
corresponding Cut in Shotgun. The Shot Updater tasks are responsible for
creating the CutItem entries. If a transcode task is running for an item in the
sequence, then it will also have access to the created CutItem in order to
update it with the newly created reviewable Version.

A new method called get_shot_parent() was exposed in the hook_get_shot hook
which is used to get (possibly creating first) an entity from SG that
corresponds to the Hiero sequence and will be the parent of Shots being
created/update in the shot updater tasks. The method is also used to retrieve
the entity link for the Cut. The default implementation returns a Sequence
entity. The new code handles the case where existing clients who have custom
versions of the hook_get_shot hook, though to make full use of the new cut
schema, they will need to add the new method. The previous version of the hook
had a method called _get_sequence() which had much the same logic, so it should
be fairly straight forward to transition to the new style hook.

A Cut Type option was added to the Shotgun Shot and Sequence Creation Settings
which populates a combo box with the valid cut type values from the site
schema. The value is saved as a preset like the existing settings in the export
UI.

Use of the collating shots options will disable the Cut and CutItem creation as
their meanings are ill defined when collating in Hiero. A new note is shown in
the exporter to make this clear.

Closes #29.