#38124 Backward compatibility fixes for Hiero 10
The primary fix here is to account for changes in behavior in Hiero 10 with
respect for collating shots.

An additional backward compatibility fix is added that changes use the newer
item.copy() if possible, else fallback to the deprecated item.clone(). This
eliminates deprecation warnings in the script editor in hiero/nuke.

Also ensures _cutHandles is defined when collating with Clip Length export
option selected.

Closes #32.