#38337 Fixes issue with preview quicktimes not being uploaded to SG
This change adds a temporary monkey patch for the Hiero method that checks
whether or not the Nuke frame server is running. We do this to allow our
preview quicktimes created in the transcode task to continue to be uploaded
when doing individual frame exports.

In previous versions of Hiero, the .nk script was always executed as a
LocalNukeRenderTask. As of Hiero 10, a new task type is used for frame exports
called FrameServerRenderTask. This task type renders frames in an individual
frame context within Nuke and therefore our quicktime write node never runs.
Thus no quicktime upload for the SG Version and no thumbnail.

By overriding this method and forcing a value of False, we trick the Hiero shot
processor internals into thinking that there is no frame server running and
cause it to use the fallback LocalNukeRenderTask which does what we need.

This fix is a work around and given time we might consider a better solution.
That might include separating the preview quicktime generation into a separate
.nk script and transcode task.

Closes #34.