#38341 Addresses regressions with Collate exports
These changes pull in the Hiero 10 collate logic into the custom exporter. The
code now calls the appropriate collate logic based on which version of Hiero is
being used. Collate was busted for old Hiero (values were offset by 1000). This
has been fixed, and these changes should be fairly backward compatible.

Most of the changes in the collating_exporter code are pulled from Hiero
itself. With the updated collate logic, we've lost the audio support when
collating that was added previously.

The other changes are in the shot_updater code and handle the various
calculations required to determine the proper head, tail, cut in, cut out
values based on what we get back from Hiero. The logic differs depending
depending on the options chosen at export time: Collate or not, Clip Length or
Cut Length, Handles, Custom start frame, etc.

Long term, it would be great to have better methods available in hiero itself
to calculate values. Even with the updated collate logic, the values returned
are inconsistent and even incorrect it seems given the various manipulations we
have to do to get the values to match what is actually written to disk.

Closes #35.