For #46368: Customizable export UI, plus additional hook methods for Shot/Cut updating. (#42)
This is implementing the idea outlined in #41 from a client. It's providing hook methods for building custom widgets for each processor we provide, plus allows for custom preset properties to be created, and those properties then being used to set/initialize the custom widget.

In addition, to meet the workflow requirements outlined by a client, additional hook methods have been added to allow for customization of if/how Shots, Cuts, and CutItems are created/updated during the export process.